Added Random Numbers
Hindsight it would have been much easier to save the list of cleaned names as a csv, then imported it, but this works. Because I coded it to save the list as a txt tile, I had to use some creative coding to get it to read the list, then turn that into a list.